feat: add image edits endpoint
