+++ litellm/proxy/common_request_processing.py
@@ -121,6 +121,7 @@ class ProxyBaseLLMRequestProcessing:
             "acancel_fine_tuning_job",
             "alist_fine_tuning_jobs",
             "aretrieve_fine_tuning_job",
+            "aimage_edit",
         ],
         version: Optional[str] = None,
         user_model: Optional[str] = None,
@@ -197,6 +198,7 @@ class ProxyBaseLLMRequestProcessing:
             "_arealtime",
             "aget_responses",
             "adelete_responses",
+            "aimage_edit",
         ],
         proxy_logging_obj: ProxyLogging,
         general_settings: dict,
